requirements.txt
Flask==2.3.3
supabase==2.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
postgrest==0.13.0  # or another version <0.14.0
